frontend/requirements.txt
streamlit>=1.30,<2
requests>=2.31,<3
pillow>=10.0,<12
streamlit-image-comparison==0.0.4